bug: debian ecosystems not correctly supported.
Debain format for the ecosystem can be of the format
Debian:3.1, which broken serde when trying to read
the vulnerability record.

Additionally the ranges field is actually optional as
practicioners might choose to only specify a specific
version or in some cases specify vesions: ["all"].

This corrects the majority of issues identified by the
smoketests.