bug: improve maven repository support
The latest version of the OSV schema allows for an optional
Maven repository to be appended to the Maven ecosystem
field. This change refactors the Ecosystems type to support
this use case and opportunistically updates the Debian and
Alpine Linux types to hold the version as an optional string
as they are following a similar convention.