schema: Support more Ubuntu quirks
Ubuntu keep adding items to their ecosystem strings, resulting in the
regex regularly failing and dropping all data. Instead, relax the regex
to allow it to parse only the start of the line. Having most of the data
is better than none. An example of this is "USN-7550-2.json".

Additionally, Ubuntu OSVs now include an ubuntu specific severity.
This otherwise matches the existing schema but is expected to contain a
simple string rather than the CVSS score. An example of this is
"UBUNTU-CVE-2025-9825.json".